# backend/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ketches api server",
        "title": "Ketches API Server",
        "contact": {},
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/envs": {
            "get": {
                "description": "List envs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Env"
                ],
                "summary": "List Envs",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "pageNo",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "projectID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ListEnvsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new env",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Env"
                ],
                "summary": "Create Env",
                "parameters": [
                    {
                        "description": "Create Env Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateEnvRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EnvModel"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/envs/refs": {
            "get": {
                "description": "Get all envs for refs under a specific cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Env"
                ],
                "summary": "All Env Refs Under Cluster",
                "parameters": [
                    {
                        "type": "string",
                        "name": "projectID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.EnvRef"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/envs/{envID}": {
            "get": {
                "description": "Get env by env ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Env"
                ],
                "summary": "Get Env",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Env ID",
                        "name": "envID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EnvModel"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing env",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Env"
                ],
                "summary": "Update Env",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Env ID",
                        "name": "envID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Env Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateEnvRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EnvModel"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an env by env ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Env"
                ],
                "summary": "Delete Env",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Env ID",
                        "name": "envID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/envs/{envID}/ref": {
            "get": {
                "description": "Get env ref by env ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Env"
                ],
                "summary": "Get Env Ref",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Env ID",
                        "name": "envID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.EnvRef"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateEnvRequest": {
            "type": "object",
            "required": [
                "clusterID",
                "displayName",
                "projectID",
                "slug"
            ],
            "properties": {
                "clusterID": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "model.EnvModel": {
            "type": "object",
            "properties": {
                "clusterID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "envID": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "model.EnvRef": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "envID": {
                    "type": "string"
                },
                "projectID": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "model.ListEnvsResponse": {
            "type": "object",
            "properties": {
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EnvModel"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.UpdateEnvRequest": {
            "type": "object",
            "required": [
                "displayName"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "envID": {
                    "type": "string"
                }
            }
        },
...
    }
}